Calculate Percentages

+++ GoalMakerServer/GoalMakerServer/DTOS/GoalDTO.cs
@@ -9,7 +9,7 @@ namespace GoalMakerServer.DTOS
     public class GoalDTO
     {
         public string Name { get; set; }
-        public int PercentageOfSuccess { get; set; }
+        public double PercentageOfSuccess { get; set; }
         public int ConfidenceLevel { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
