+++ GoalMakerServer/GoalMakerServer/DTOS/TeamDTO.cs
@@ -12,7 +12,7 @@ namespace GoalMakerServer.DTOS
         public string Name { get; set; }
         public int OrganizationId { get; set; }
         public User Owner { get; set; }
-        public int PercentageOfSuccess { get; set; }
+        public double PercentageOfSuccess { get; set; }
         public int ConfidenceLevel { get; set; }
         public string TeamCountry { get; set; }
